initial wip

+++ tool/tctl/common/resources/scoped_role.go
@@ -27,6 +27,7 @@ import (
 	"github.com/gravitational/teleport/lib/asciitable"
 	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/itertools/stream"
+	"github.com/gravitational/teleport/lib/scopes"
 	scopedaccess "github.com/gravitational/teleport/lib/scopes/access"
 	scopedutils "github.com/gravitational/teleport/lib/scopes/utils"
 	"github.com/gravitational/teleport/lib/services"
@@ -66,8 +67,8 @@ func (c *scopedRoleCollection) WriteText(w io.Writer, verbose bool) error {
 	return trace.Wrap(err)
 }
 
-func scopedRoleHandler() Handler {
-	return Handler{
+func scopedRoleScopedHandler() ScopedHandler {
+	return ScopedHandler{
 		getHandler:    getScopedRole,
 		createHandler: createScopedRole,
 		updateHandler: updateScopedRole,
@@ -133,15 +134,21 @@ func updateScopedRole(ctx context.Context, client *authclient.Client, raw servic
 	return nil
 }
 
-func getScopedRole(ctx context.Context, client *authclient.Client, ref services.Ref, opts GetOpts) (Collection, error) {
-	if ref.Name != "" {
+func getScopedRole(ctx context.Context, client *authclient.Client, subKind string, sqn *scopes.QualifiedName, opts GetOpts) (Collection, error) {
+	if subKind != "" {
+		return nil, rejectSubKind(scopedaccess.KindScopedRole, subKind)
+	}
+
+	if sqn != nil {
 		rsp, err := client.ScopedAccessServiceClient().GetScopedRole(ctx, &scopedaccessv1.GetScopedRoleRequest{
-			Name: ref.Name,
+			Name: sqn.Name,
 		})
 		if err != nil {
 			return nil, trace.Wrap(err)
 		}
-
+		if rsp.Role.GetScope() != sqn.Scope {
+			return nil, scopeMismatchNotFound(scopedaccess.KindScopedRole, *sqn, rsp.Role.GetScope())
+		}
 		return &scopedRoleCollection{roles: []*scopedaccessv1.ScopedRole{rsp.Role}}, nil
 	}
 
@@ -152,16 +159,31 @@ func getScopedRole(ctx context.Context, client *authclient.Client, ref services.
 	return &scopedRoleCollection{roles: items}, nil
 }
 
-func deleteScopedRole(ctx context.Context, client *authclient.Client, ref services.Ref) error {
+func deleteScopedRole(ctx context.Context, client *authclient.Client, subKind string, sqn scopes.QualifiedName) error {
+	if subKind != "" {
+		return rejectSubKind(scopedaccess.KindScopedRole, subKind)
+	}
+
+	// Fetch first to verify scope before deleting.
+	rsp, err := client.ScopedAccessServiceClient().GetScopedRole(ctx, &scopedaccessv1.GetScopedRoleRequest{
+		Name: sqn.Name,
+	})
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	if rsp.Role.GetScope() != sqn.Scope {
+		return scopeMismatchNotFound(scopedaccess.KindScopedRole, sqn, rsp.Role.GetScope())
+	}
+
 	if _, err := client.ScopedAccessServiceClient().DeleteScopedRole(ctx, &scopedaccessv1.DeleteScopedRoleRequest{
-		Name: ref.Name,
+		Name: sqn.Name,
 	}); err != nil {
 		return trace.Wrap(err)
 	}
 	fmt.Printf(
 		"%v %q has been deleted\n",
 		scopedaccess.KindScopedRole,
-		ref.Name,
+		sqn.Name,
 	)
 	return nil
 }
